[CMake/Bazel] Support usage of opt driver as a library
