+++ llvm/tools/opt/CMakeLists.txt
@@ -29,12 +29,24 @@ set(LLVM_LINK_COMPONENTS
   Passes
   )
 
-add_llvm_tool(opt
+# We don't want to link this into libLLVM
+add_llvm_library(LLVMOptDriver
+  STATIC
   NewPMDriver.cpp
-  opt.cpp
+  optdriver.cpp
+  PARTIAL_SOURCES_INTENDED
+  DEPENDS
+  intrinsics_gen
+)
 
+add_llvm_tool(opt
+  PARTIAL_SOURCES_INTENDED
+  opt.cpp 
   DEPENDS
   intrinsics_gen
   SUPPORT_PLUGINS
+
   )
+target_link_libraries(opt PRIVATE LLVMOptDriver)
+
 export_executable_symbols_for_plugins(opt)
