+++ utils/bazel/llvm-project-overlay/llvm/BUILD.bazel
@@ -4812,8 +4812,8 @@ cc_binary(
     ],
 )
 
-cc_binary(
-    name = "opt",
+cc_library(
+    name = "opt-driver",
     srcs = glob([
         "tools/opt/*.cpp",
         "tools/opt/*.h",
@@ -4824,7 +4824,6 @@ cc_binary(
         "@platforms//os:macos": [],
         "//conditions:default": ["-Wl,--export-dynamic"],
     }),
-    stamp = 0,
     deps = [
         ":AllTargetsAsmParsers",
         ":AllTargetsCodeGens",
@@ -4849,6 +4848,12 @@ cc_binary(
         ":common_transforms",
         ":config",
     ],
+)
+
+cc_binary(
+    name = "opt",
+    stamp = 0,
+    deps = [":opt-driver"]
 )
 
 gentbl(
